ListsTools: inline ListsTools2
